+++ src/Util.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <cstdint>
+
 class Util {
   public:
     static void flush();
-    static int get_ms();
+    static uint64_t get_ms();
     static int str2move(const char move_old[6]);
     static char *move2str(char *move_str, const int move);
 };
